Move provider files in provider directory
