+++ gptme/chat.py
@@ -15,7 +15,7 @@ from .interrupt import clear_interruptible, set_interruptible
 from .llm import reply
 from .logmanager import Log, LogManager, prepare_messages
 from .message import Message
-from .models import get_model
+from .providers.models import get_model
 from .prompts import get_workspace_prompt
 from .readline import add_history
 from .tools import ToolUse, execute_msg, has_tool
