+++ gptme/init.py
@@ -6,7 +6,7 @@ from rich.logging import RichHandler
 
 from .config import config_path, get_config, set_config_value
 from .llm import init_llm
-from .models import (
+from .providers.models import (
     PROVIDERS,
     Provider,
     get_recommended_model,
