+++ gptme/providers/llm_openai_models.py
