+++ gptme/reduce.py
@@ -9,7 +9,7 @@ from collections.abc import Generator
 
 from .codeblock import Codeblock
 from .message import Message, len_tokens
-from .models import get_model
+from .providers.models import get_model
 
 logger = logging.getLogger(__name__)
