+++ tests/test_server.py
@@ -9,7 +9,7 @@ flask = pytest.importorskip(
 # noreorder
 from flask.testing import FlaskClient  # fmt: skip
 from gptme.init import init  # fmt: skip
-from gptme.models import get_model  # fmt: skip
+from gptme.providers.models import get_model  # fmt: skip
 from gptme.server.api import create_app  # fmt: skip
